+++ src/App.vue
@@ -1,21 +1,23 @@
 <template>
   <div id="app">
     <div id="nav">
+      <TheNav/>
       <b-card title="Card Title" no-body>
-        <b-card-header header-tag="nav">
-          <b-nav card-header pills>
-            <b-nav-item to="/" exact exact-active-class="active" >ZHome</b-nav-item>
-            <b-nav-item to="/config" exact exact-active-class="active">Pengaturan</b-nav-item>
-            <b-nav-item to="/route/list" exact exact-active-class="active">Rute</b-nav-item>
-          </b-nav>
-        </b-card-header>
         <b-card-body class="text-center">
-         <router-view/>
+         <router-view v-bind:key="$route.path"/>
         </b-card-body>
       </b-card> 
     </div>  
   </div>
 </template>
 
+<script>
+import TheNav from '@/components/TheNav'
 
+export default {
+  components: {
+    TheNav
+  }
+}
+</script>
