--- src/db/weight2.json
@@ -1,1 +1,0 @@
-["tomato"]
