+++ src/views/PageRoute.vue
@@ -3,20 +3,16 @@
   <b-card no-body>
     <b-card-header header-tag="nav">
       <b-nav card-header tabs>
-        <b-nav-item to="/route/list">Rute</b-nav-item>
-        <b-nav-item to="/route/busstop">Bus Stop</b-nav-item>
+        <b-nav-item><router-link v-bind:to="{name: 'RouteList'}">Rute</router-link></b-nav-item>
+        <b-nav-item><router-link v-bind:to="{name: 'RouteBusStop'}">Bus Stop</router-link></b-nav-item>
       </b-nav>
     </b-card-header>
-      <router-view/>
+      <router-view v-bind:key="$route.path"/>
     <b-card-body >
     </b-card-body>
   </b-card>
 </div>
 </template>
-
-
-
-
 
 <script>
 export default {
